+++ packages/common/web/src/types/editorTypes.ts
@@ -62,11 +62,9 @@ export const DOC_STYLE_TYPES = {
   H4: 'headerFour',
   H5: 'headerFive',
   H6: 'headerSix',
-  OL: 'orderedList',
-  UL: 'unorderedList',
 };
 
-export const DRAFT_TO_RICOS_DOC_TYPE = {
+export const DRAFT_TO_DOC_TYPE = {
   [HEADER_BLOCK.ONE]: 'headerOne',
   [HEADER_BLOCK.TWO]: 'headerTwo',
   [HEADER_BLOCK.THREE]: 'headerThree',
@@ -76,8 +74,8 @@ export const DRAFT_TO_RICOS_DOC_TYPE = {
   [HEADER_BLOCK.PARAGRAPH]: 'paragraph',
 };
 
-export const DRAFT_TO_DOC_TYPE = {
-  ...DRAFT_TO_RICOS_DOC_TYPE,
-  'ordered-list-item': 'orderedList',
-  'unordered-list-item': 'unorderedList',
+export const DRAFT_TO_DOC_TYPE_WITH_LISTS = {
+  ...DRAFT_TO_DOC_TYPE,
+  'ordered-list-item': 'paragraph',
+  'unordered-list-item': 'paragraph',
 };