+++ packages/editor/web/src/RichContentEditor/utils/editorCommandsUtils.ts
@@ -28,9 +28,8 @@ import {
   dynamicStyleParsers,
   safeJsonParse,
   draftDecorationsToCss,
-  DRAFT_TO_RICOS_DOC_TYPE,
   DRAFT_TO_DOC_TYPE,
-  DOC_STYLE_TYPES,
+  DRAFT_TO_DOC_TYPE_WITH_LISTS,
 } from 'wix-rich-content-common';
 
 import { DRAFT_TO_RICOS_CUSTOM_STYLES, defaultFontSizes, defaultMobileFontSizes } from './consts';
@@ -41,7 +40,7 @@ export const getWiredFontStyles = (
   isMobile?: boolean
 ) => {
   const fontSizes = {};
-  Object.entries(DRAFT_TO_RICOS_DOC_TYPE).forEach(([draftHeader, ricosHeader]) => {
+  Object.entries(DRAFT_TO_DOC_TYPE).forEach(([draftHeader, ricosHeader]) => {
     fontSizes[ricosHeader] = {
       'font-size':
         documentStyle?.[ricosHeader]?.['font-size'] ||
@@ -135,11 +134,7 @@ export const getFontSize = (editorState: EditorState) => {
 const getBlockStyle = (editorState: EditorState, getDocumentStyle) => {
   const blockType = getBlockType(editorState);
   const documentStyle = getDocumentStyle?.();
-  let docStyleType = DRAFT_TO_DOC_TYPE[blockType];
-  docStyleType = [DOC_STYLE_TYPES.OL, DOC_STYLE_TYPES.UL].includes(docStyleType)
-    ? DOC_STYLE_TYPES.P
-    : docStyleType;
-  return documentStyle?.[docStyleType];
+  return documentStyle?.[DRAFT_TO_DOC_TYPE_WITH_LISTS[blockType]];
 };
 
 export const setFontSize = (