+++ packages/editor/web/src/RichContentEditor/EditorCommands.ts
@@ -26,6 +26,7 @@ import {
   getAnchorableBlocks,
   removeCurrentInlineStyle,
   isAtomicBlockInSelection,
+  isTextBlockInSelection,
   scrollToBlock,
   insertCustomLink,
   getSelectedBlocks,
@@ -166,6 +167,7 @@ export const createEditorCommands = (
     getAnchorBlockInlineStyles: EditorCommands['getAnchorBlockInlineStyles'];
     getWiredFontStyles: EditorCommands['getWiredFontStyles'];
     isAtomicBlockInSelection: EditorCommands['isAtomicBlockInSelection'];
+    isTextBlockInSelection: EditorCommands['isTextBlockInSelection'];
     getAnchorBlockType: EditorCommands['getAnchorBlockType'];
   } = {
     getSelection: () => {
@@ -229,6 +231,7 @@ export const createEditorCommands = (
       return blocks.some(block => block.getKey() === blockKey);
     },
     isAtomicBlockInSelection: () => isAtomicBlockInSelection(getEditorState()),
+    isTextBlockInSelection: () => isTextBlockInSelection(getEditorState()),
     getAnchorBlockType: () => getBlockType(getEditorState()),
   };