+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/core-configs.ts
@@ -1,4 +1,5 @@
 import { RicosExtension } from 'ricos-tiptap-types';
+import { createCommandsConfig } from '../../core-commands';
 import {
   createDraftConfig,
   createFocusConfig,
@@ -13,4 +14,5 @@ export const coreConfigs: RicosExtension[] = [
   createStylesConfig(),
   createFocusConfig(),
   createUniqueId(),
+  createCommandsConfig(),
 ];